Initial code for reading new revlog cols from db.
This adds a backend command `get_revlog_entry(revlog_id)`, implemented
in a new backend `Revlog` service, which returns a `revlog_entry` object
corresponding to a row of the `revlog` db table.  The `revlog_entry`
includes fields corresponding to `feedback`, `tags`, and `mod` columns
of the table.

This also adds a test of the command in `pylib/tests/test_revlog.py`.