Hack to sync augmented RevlogEntry to Anki servers.
The Anki servers reject the serialized `revlog::RevlogEntry` object because it
has been augmented with three additional fields that aren't in the
official Anki schema for the `revlog` table.

This adds a local `sync::RevlogEntry` object to the sync code which
lacks the additional fields and is therefore compatible with the
official Anki server and database schema. The new code copies back and
forth between `revlog::RevlogEntry` and `sync::RevlogEntry`, omitting
the additional incompatible fields.

At some point I'll want a local sync server that supports the additional
fields. To make it compatible with standard Anki, I'll need to make sure
that it syncs with both versions of `RevlogEntry`.