Add support for editing review feedback and tags.
This required some hacks to reenable syncing to the official Anki
server, which doesn't support the three fields I've added to the
`revlog` table.

At some point I'll want to make a local sync server that does handle
these fields.

I'll also want to add support for editing additional `revlog`
information, such as the time taken to review the card, the date/time of
the review, and the ease button that was pressed. I think this will
entail modifying some scheduler code in order to maintain a consistent
(rewritten) scheduler history. (Changing the review date and ease
will amount to rewriting the review history for the card.)