Add reviews mode to browser.
This adds a reviews mode to the browser, and changes from a notes/cards
mode switch to a set of three buttons, N/C/R, for notes, cards, and
reviews modes, respectively.

I've changed the behavior upon clicking a tag in the sidebar tag-tree.
Now instead of entering "tag:..." in the search bar, it enters "tag:...
OR rtag:..." in order to search both notes and reviews.

This commit also has an initial change to the tags-tree editor in the
side bar, so that when a tag is deleted, it is removed from reviews in
addition to the usual removal from notes. This is demo code in order to
work out some of the required architecture changes. More work is needed,
some described in the to-do list below.

To do:
- Add code to handle reviews during reparenting, renaming, and
  find-and-replace of tags. I'll also want new functionality for adding
  and removing tags to/from reviews.
- Add more translations to recent code.
- I'm not entirely confident of the code in the new browser review mode
  and associated search code, so I'll want to add more tests.