Embed Jupyter kernel.
This embeds a Jupyter kernel in Anki. It's running in its own thread,
which makes race conditions possible, so I want to switch to asyncio
ASAP. This kernel is for development purposes; I personally like having
it, but most people wouldn't have any use for it, so I'll probably
remove it at some point.