Mod scheduler to answer card at given timestamp.
This replaces several Rust functions in the v3 scheduler, each of which
uses `TimestampSecs::now()`, with versions that use
`TimestampSecs::now()` by default but optionally allow specifying a
different timestamp. If the original function was named `fubar(...)`
then the new version consists of `fubar_at(...)` which accepts a given
timestamp, and a rewritten `fubar(...)` which calls `fubar_at(...)` with
the result of `TimestampSecs::now()`.

I frequently study without a device handy for running Anki, in which
case I record by hand what problems I studied, when, how I answered, and
how long it took. I want to record this information after the fact.
While Anki's v3 Anki scheduler allows specifying a timestamp when
answering a card (the `CardAnswer` object accepts an
`answered_at_millis` field), the resulting card state is not the same as
it would have been were `answered_at_millis` the same as `now` (the
current Unix timestamp times 1000). In part this is because several of
the scheduling functions use `TimestampSecs::now()` instead of the given
timestamp.

The changes is this commit propagate timestamps based on
`answered_at_millis` throughout the call tree when answering a card.
This gives the expected result when the given timestamp for the new
review is later than that of the most recent previous review for the
card.

A separate kind of inconsistency results when the given timestamp is
_prior_ to that of the most recent review. In other words, the card's
state is inconsistent because the review history has been rewritten, and
the card's state has been updated in a different order than its new
history. This kind of problem will be addressed shortly.