Add scheduler/backend functions to replay reviews.
This commit provides the basis for replaying a card's review history.

If a card answer is given with a timestamp that is earlier than that of
the most recent previous answer, the resulting card state is
inconsistent. This is because the card's state changes have taken place
out of order. The v3 scheduler was not designed to handle this situation.

This commit adds a pair of functions to the scheduler named
`replay_card_histories(...)` and `replay_card_histories_inner(...)`. The
former calls the latter after setting up an undo transaction.

This code is messy and not well tested. I'm committing it now mainly to
save it. I plan to give it some cleanup and better testing.