Backend changes to support deleting reviews.
This dovetails with the recent database refactoring in order to try to
maintain compatibility with Anki servers, which don't support deleting
reviews from the `revlog` table. To achieve this I hide rather than
delete any reviews that have been synced with Anki servers. The hiding
is done via the `reviews` database view. The logic for handling this
also lives in the database. More info about the goals of this refactor
are in the previous AnkiMath commit.