Frontend changes to support deleting reviews.
This allows deleting reviews in from the browser in "reviews" mode. It
supports undo. In this commit I've also moved the browser "node-delete"
command to the "edit" menu, and renamed it to just "delete", so it can
be used to either delete notes from the "notes" or "cards" modes, or
reviews from the "reviews" mode. I've tried to keep these operations
compatible with the Anki servers, which don't support deleting reviews.
More info about this can be found in the two previous AnkiMath commits.