Fix notes not in searches if they lack reviews.
I created a bug where a notes and cards weren't appearing in certain
searches because they hadn't been reviewed yet. This was caused by using
a pair of inner joins from notes to cards and cards to reviews, which
results in an empty query when there are no reviews yet. I solved this
by switching to a left join between cards and reviews, which allows for
null reviews to appear in the query results.

This required changing Rust function `add_regexp_function`, which is
used to add a `regexp` SQLite function for Rusqlite, so that `regexp`
gracefully handles null values resulting from the left join. The desired
behavior in this case is for `regexp` to return `false`.