Hack to sync with official Anki servers.
This is temporary. It replaces the structure `revlog::RevlogEntry` with
a local version version compatible with the official Anki servers. The
hack is confined to the file `rslib/src/sync/collection/chunks.rs`.

I'll want to replace this hack with code that senses whether it's
dealing with standard Anki servers/clients or with AnkiMath
servers/clients, and syncs differently depending on which version it
finds.