Parse/propagate server version info to sync code.
This adds a `server_version` field to Rust `struct ClientSyncState`.
This will be used in syncing code to sync additional AnkiMath `revlog`
fields between AnkiMath client and server, but only sync standard
`revlog` fields for non-AnkiMath client or server.

This `server_version` field is populated with data sent by the server in
an HTTP `Server` header in `SyncResponse::make_response()` defined in
`rslib/src/sync/response.rs`, and extracted by the client in
`IoMonitor::zstd_request_with_timeout()` defined in
`rslib/src/sync/http_client/io_monitor.rs`.

This is not a permanent solution. A better place for this information
would be HTTP sessions. It looks like Anki development is headed in this
direction. But until it arrives, the above solution is more in tune with
Anki's current stateful syncing, with state stored in and bouncing back
and forth between the client and the server across multiple HTTP
transactions.

I'll soon add similar code to store a `client_version` field in the
`struct ServerSyncState`.