AnkiMath addon providing random seed.
Provides a card field `{{randseed-Seed:}}` unique to each card, that
changes when the user presses a "success" button (e.g. easy, good,
or hard -- which buttons are considered "successful" is configurable).

This is to allow varying a math problem after the user solves it, but not
before. For example, a card may ask the user to solve

  ```latex
  \[
    \int_{1}^{2} w^2 \ln w \,dw
  \]
  ```

and once solved, the card might change, swapping 3 and 4 in place of 1 and 2.
This is done by placing code like the following in the `Front` template for the
card:

 ```javascript
 <script>
   if (typeof random_seed === 'undefined') { let random_seed; }
   if (`{{randseed-Seed:}}`.startsWith('Unknown field:')) {
     random_seed = 0;
   } else {
     random_seed = {{randseed-Seed:}};
   }

   {{Params}}

   document.getElementById('parameterized_front').innerHTML = String.raw`{{Front}}`;
 </script>

 <div id='parameterized_front'>
   (Warning: Front not yet defined by card template script.)
 </div>
 ```

And similar code in the `Back` template. The card should have a `Params` field
reading something like:

  ```javascript
  if (typeof params === 'undefined') { let params, a, b; }
  params = [
    [1,2],
    [3,4],
    [2,3],
  ];
  [a, b] = params[random_seed % params.length];
  ```

And a `Front` field like:

  ```latex
  \[
  \int_{${a}}^{${b}} w^2 \ln w \,dw
  \]
  ```

Similarly for the `Back` field, where `a` and `b` are randomly selected in the
`Params` field and then interpolated.

Problem to solve: doing this causes a delay displaying the card while the
database is searched for the last successful review.