In review heatmap, weight reviews by button chosen.
When computing "trouble" scores for cards, previously all reviews were
given the same weight. Now the original weight is used for reviews with
the "again" button clicked, but is decreased by 25% when "hard" was
clicked, by 50% for "good", and by 75% for "easy". I'll eventually want
to make this configurable, once I move this heatmap idea out of the
notebook and into the GUI.