Brainstorm interactive review heatmaps.
This brainstorm uses plotly/dash, dash_bio, and jupyter_dash to make
interactive clustered heatmaps with dendrograms along the x-axis
(clustering review tags) and the y-axis (note tags). The brainstorm is
in *230208-1401-dash-bio-clustergram.ipynb*, the metric used being the
one devised in *_230117-1908 display stats.ipynb*.

In fact, since both notebooks run the same Anki kernel, I'm generating
the metrics in a dataframe in *_230117-1908 display stats.ipynb*, and
then using the same dataframe in
*230208-1401-dash-bio-clustergram.ipynb* to generate the interactive
heatmap. Working in two separate notebooks on the same kernel is bad
form, but this is just a brainstorm that I plan to move into Anki
proper.