Added a "topic_cards" addon.
The purpose of this addon is "topic cards" that exercise recalling
skills rather than facts. Each topic card specifies an Anki deck from
which to draw flashcards with math problems on a particular topic, such
as integration using partial fractions. The goal is not to memorize the
math problems, but to memorize the skill. Each time the "integration
with partial fractions" topic card is shown, it pulls a new math problem
from the flashcard deck. Both the topic deck and the flashcard deck have
their own scheduling queues.

This addon is very rough. I need to work on the following issues:

- Undo is messy. This is because while the topic card is being rendered,
  the addon switches active decks between the topic-card deck and the
  flash-card deck in order to access the flash-card deck's queue for
  scheduling. Each time a deck switch takes place, an entry for that
  operation is placed onto the undo stack. So undoing a review of a
  topic deck requires undoing a deck switch, then a flash-card review,
  then another deck switch, then the topic-card review, and so on.

- Need to handle the case of a topic-card for a flash-card deck that's
  all caught up (or empty). I tried autmatically burying the topic-card,
  but then when several topic-cards are buried in rapid succession (when
  their flash-card decks are all empty), there's a race condition
  resulting in the reviewer losing its reference to the topic-card. So
  I've disabled that solution attempt.